techpack: disp: msm: sde: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I7e063738d3078ffbf29076c887b818f1d1b76c13
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>